Creating CalcuationResponseModel in tests, updating fields in models

+++ app/models/hip/calculation/EndOfYearEstimate.scala
@@ -19,7 +19,7 @@ package models.hip.calculation
 import play.api.libs.json._
 
 case class EndOfYearEstimate(incomeSource: Option[List[IncomeSource]],
-                             totalAllowancesAndDeductions: Option[Int] = None,
+                             totalAllowancesAndDeductions: Option[BigDecimal] = None,
                              totalEstimatedIncome: Option[Int] = None,
                              totalTaxableIncome: Option[Int] = None,
                              incomeTaxAmount: Option[BigDecimal] = None,
